backend/requirements.txt
django
channels
daphne
opencv-contrib-python
numpy
google-genai
python-dotenv
